Renaming Files
changed files that start with small letter to capital to follow one naming convention once and for all
